refactor notebook button
