+++ js_modules/dagit/packages/core/src/metadata/MetadataEntry.tsx
@@ -5,7 +5,6 @@ import {
   Colors,
   DialogFooter,
   Dialog,
-  DialogBody,
   Group,
   Icon,
   Markdown,
@@ -17,11 +16,11 @@ import * as React from 'react';
 import {Link} from 'react-router-dom';
 import styled from 'styled-components/macro';
 
-import {AppContext} from '../app/AppContext';
 import {copyValue} from '../app/DomUtils';
 import {assertUnreachable} from '../app/Util';
 import {displayNameForAssetKey} from '../asset-graph/Utils';
 import {assetDetailsPathForKey} from '../assets/assetDetailsPathForKey';
+import {NotebookButton} from '../ui/NotebookButton';
 
 import {TableSchema, TABLE_SCHEMA_FRAGMENT} from './TableSchema';
 import {MetadataEntryFragment} from './types/MetadataEntryFragment';
@@ -76,14 +75,6 @@ export const MetadataEntry: React.FC<{
   expandSmallValues?: boolean;
   repoLocation?: string;
 }> = ({entry, expandSmallValues, repoLocation}) => {
-  const {rootServerURI} = React.useContext(AppContext);
-  const [open, setOpen] = React.useState(false);
-  React.useEffect(() => {
-    const onOpen = () => setOpen(true);
-    document.addEventListener('show-kind-info', onOpen);
-    return () => document.removeEventListener('show-kind-info', onOpen);
-  }, []);
-
   switch (entry.__typename) {
     case 'PathMetadataEntry':
       return (
@@ -184,38 +175,7 @@ export const MetadataEntry: React.FC<{
       return <TableSchema schema={entry.schema} />;
     case 'NotebookMetadataEntry':
       if (repoLocation) {
-        return (
-          <div>
-            <Button icon={<Icon name="content_copy" />} onClick={() => setOpen(true)}>
-              View Notebook
-            </Button>
-            <Dialog
-              icon="info"
-              onClose={() => setOpen(false)}
-              style={{width: '80vw', maxWidth: 900}}
-              title={entry.path.split('/').pop()}
-              usePortal={true}
-              isOpen={open}
-            >
-              <DialogBody>
-                <iframe
-                  title={entry.path}
-                  src={`${rootServerURI}/dagit/notebook?path=${encodeURIComponent(
-                    entry.path,
-                  )}&repoLocName=${repoLocation}`}
-                  sandbox=""
-                  style={{border: 0, background: 'white'}}
-                  seamless={true}
-                  width="100%"
-                  height={500}
-                />
-              </DialogBody>
-              <DialogFooter>
-                <Button onClick={() => setOpen(false)}>Close</Button>
-              </DialogFooter>
-            </Dialog>
-          </div>
-        );
+        return <NotebookButton path={entry.path} repoLocation={repoLocation} />;
       }
       return (
         <Group direction="row" spacing={8} alignItems="center">
